+++ lib/persistor/sqlite_persistor/sqlite_data_store_config.dart
@@ -32,7 +32,7 @@ class SqliteDataStoreConfig extends DataStoreConfig {
             final value = rows.first[_valueColumn] as String;
 
             final json =
-                jsonDecode(encrypted ? encrypter.decrypt(value) : value);
+                jsonDecode(encrypted ? await encrypter.decrypt(value) : value);
             final store = ValueStore<ValueStore>();
 
             for (final entry in json.entries) {
@@ -50,7 +50,8 @@ class SqliteDataStoreConfig extends DataStoreConfig {
               _tableName,
               {
                 _keyColumn: name,
-                _valueColumn: encrypted ? encrypter.encrypt(value) : value,
+                _valueColumn:
+                    encrypted ? await encrypter.encrypt(value) : value,
               },
               conflictAlgorithm: ConflictAlgorithm.replace,
             );