+++ test/web/indexed_db_persistor_test.dart
@@ -12,9 +12,9 @@ void main() {
     getStore: (
       persistor,
       storeName, {
-      bool encrypted = false,
+      required encrypted,
     }) async {
-      final result = await persistor.runTransaction('Get', (objectStore) {
+      final result = await persistor.runTransaction('Get', (objectStore) async {
         return objectStore.get(storeName.toJS);
       });
 
@@ -24,7 +24,9 @@ void main() {
 
       final value = result[IndexedDBPersistor.valuePath];
 
-      return jsonDecode(encrypted ? persistor.encrypter.decrypt(value) : value);
+      return jsonDecode(
+        encrypted ? await persistor.encrypter.decrypt(value) : value,
+      );
     },
     factory: IndexedDBPersistor.new,
   );